requirements.txt
groq==0.4.1
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
